+++ src/main/java/kitchenpos/order/application/TableService.java
@@ -1,10 +1,10 @@
-package kitchenpos.order_table.application;
+package kitchenpos.order.application;
 
 import java.util.List;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.repository.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
