+++ src/main/java/kitchenpos/order/infrastructure/persistence/JdbcTemplateOrderTableDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.order_table.infrastructure.persistence;
+package kitchenpos.order.infrastructure.persistence;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
