+++ src/main/java/kitchenpos/table_group/application/TableGroupService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTables;
-import kitchenpos.order_table.domain.repository.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTables;
+import kitchenpos.order.domain.repository.OrderTableRepository;
 import kitchenpos.table_group.domain.TableGroup;
 import kitchenpos.table_group.domain.repository.TableGroupRepository;
 import org.springframework.stereotype.Service;
