+++ src/main/java/kitchenpos/table_group/domain/TableGroup.java
@@ -2,7 +2,7 @@ package kitchenpos.table_group.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.order_table.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public class TableGroup {
