+++ src/main/java/kitchenpos/table_group/infrastructure/persistence/TableGroupRepositoryImpl.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.infrastructure.persistence.OrderTableDao;
-import kitchenpos.order_table.infrastructure.persistence.OrderTableEntity;
-import kitchenpos.order_table.infrastructure.persistence.OrderTableMapper;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.infrastructure.persistence.OrderTableDao;
+import kitchenpos.order.infrastructure.persistence.OrderTableEntity;
+import kitchenpos.order.infrastructure.persistence.OrderTableMapper;
 import kitchenpos.table_group.domain.TableGroup;
 import kitchenpos.table_group.domain.repository.TableGroupRepository;
 import org.springframework.stereotype.Repository;
