+++ src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -2,7 +2,7 @@ package kitchenpos.fixture;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.order_table.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.table_group.domain.TableGroup;
 
 public class TableGroupFixture {
