+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -21,8 +21,8 @@ import kitchenpos.menu.infrastructure.persistence.MenuGroupRepositoryImpl;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.infrastructure.persistence.OrderRepositoryImpl;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.infrastructure.persistence.OrderTableRepositoryImpl;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.infrastructure.persistence.OrderTableRepositoryImpl;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.infrastructure.persistence.ProductRepositoryImpl;
 import kitchenpos.support.ServiceIntegrationTest;
