+++ src/test/java/kitchenpos/order/domain/OrderTablesTest.java
@@ -1,9 +1,11 @@
-package kitchenpos.order_table.domain;
+package kitchenpos.order.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import java.util.stream.Stream;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTables;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
