+++ src/test/java/kitchenpos/order/infrastructure/persistence/OrderRepositoryImplTest.java
@@ -23,8 +23,7 @@ import kitchenpos.menu.infrastructure.persistence.MenuGroupRepositoryImpl;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.infrastructure.persistence.OrderTableRepositoryImpl;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.infrastructure.persistence.ProductRepositoryImpl;
 import kitchenpos.support.JdbcTestHelper;
