+++ src/test/java/kitchenpos/table_group/infrastructure/persistence/TableGroupRepositoryImplTest.java
@@ -11,8 +11,8 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import kitchenpos.fixture.OrderTableFixture;
 import kitchenpos.fixture.TableGroupFixture;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.infrastructure.persistence.OrderTableRepositoryImpl;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.infrastructure.persistence.OrderTableRepositoryImpl;
 import kitchenpos.support.JdbcTestHelper;
 import kitchenpos.table_group.domain.TableGroup;
 import org.assertj.core.api.Assertions;
